Reacommodating utils functions
