+++ src/utils/index.js
@@ -1,6 +1,5 @@
-export * from "./dateDelta.js";
 export * from "./identifyStations.js";
-export * from "./parseDate.js";
+export * from "./numeric.js";
 export * from "./passwordValidation.js";
 export * from "./plot.js";
 export * from "./stationStats.js";
